+++ lib/rubocop/cop/performance/map_compact.rb
@@ -66,7 +66,7 @@ module RuboCop
           chained_method = compact_node.parent
           compact_method_range = compact_node.loc.selector
 
-          if compact_node.multiline? && chained_method&.loc.respond_to?(:selector) &&
+          if compact_node.multiline? && chained_method&.loc.respond_to?(:selector) && chained_method.dot? &&
              !invoke_method_after_map_compact_on_same_line?(compact_node, chained_method)
             compact_method_range = range_by_whole_lines(compact_method_range, include_final_newline: true)
           else